requirements.txt
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
binwalk==2.3.3
bitstring==3.1.9
blspy==1.0.16
btrfsutil==5.18.1
CacheControl==0.12.6
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
chia-blockchain==1.7.0
chia_rs==0.2.0
chiabip158==1.2
chiapos==1.0.11
chiavdf==1.0.8
click==8.1.3
clvm==0.9.7
clvm-tools==0.4.6
clvm_tools_rs==0.1.30
colorama==0.4.5
colorlog==6.7.0
commonmark==0.9.1
concurrent-log-handler==0.9.20
cryptography==38.0.3
cycler==0.11.0
distlib==0.3.4
distro==1.7.0
dnslib==0.9.23
dnspython==2.2.1
docutils==0.18.1
filelock==3.8.0
fonttools==4.33.3
frozenlist==1.3.3
future==0.18.2
html5lib==1.1
idna==3.3
jaraco.classes==3.2.3
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
jeepney==0.8.0
jsonschema==4.6.2
keyring==23.9.3
kiwisolver==1.4.3
matplotlib==3.5.2
more-itertools==8.13.0
msgpack==1.0.3
multidict==6.0.4
nspektr==0.4.0
numpy==1.23.0
ordered-set==4.0.2
packaging==21.3
pandas==1.4.3
pep517==0.12.0
Pillow==9.2.0
platformdirs==2.5.1
ply==3.11
portalocker==2.7.0
psutil==5.9.4
pycairo==1.21.0
pycparser==2.21
Pygments==2.12.0
PyGObject==3.42.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
Reflector==2021.11.20.2.41.3
requests==2.27.1
resolvelib==0.8.1
rich @ file:///build/python-rich/src/rich-12.4.4/dist/rich-12.4.4-py3-none-any.whl
scipy==1.8.1
screenkey==1.5
SecretStorage==3.3.3
setproctitle==1.2.3
six==1.16.0
sortedcontainers==2.4.0
tenacity==8.0.1
tomli==2.0.1
trove-classifiers==2022.6.26
typing_extensions==4.4.0
urllib3==1.26.9
validate-pyproject==0.9
watchdog==2.2.0
webencodings==0.5.1
yarl==1.8.2
zstd==1.5.2.6
